// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/pageRecordsParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */

export type PageRecordsParams = {
  /**
   * 路由主键
   */
  routeId?: string;
  /**
   * 文件主键
   */
  transferId?: string;
  /**
   * 目标编码
   */
  targetCode?: string;
  /**
   * 执行状态
   */
  executeStatus?: string;
  /**
   * 页码
   */
  pageIndex?: number;
  /**
   * 每页条数
   */
  pageSize?: number;
};